Update dockerfile to use speed up build time
Instead of installing prolog through apt-get it copies the compiled
program out of the official prolog docker container

Base off of the python image which is the recommended image to use for
Django

Move some of the installs around so it takes advantage of caching
better.

Remove some of the git clones in particular for sCASP since you are
always downloading a specific commit it should be pretty safe to cache
forever.

Replace ADD with COPY so you don't blow away the cache and re-run pip
install every single time you build, I'm not sure your dependencies are
changing that much.

Change workdir of install and some paths so that everything works again

Replace python3 with python since we know python is python v3 and we
aren't using ubunut anymore.